organize ai functionality into a sub module
