--- src/datagrunt/ai/__init__.py
@@ -1,5 +1,0 @@
-"""Initialize the AI module."""
-
-from datagrunt.ai.factories import AIEngineFactory
-
-__all__ = ['AIEngineFactory']
